Conflicts:
	packages/ForTrilinos/src/amesos/foramesos.F90
	packages/ForTrilinos/src/aztecoo/foraztecoo.F90
	packages/ForTrilinos/src/epetra/forepetra.F90
	packages/ForTrilinos/src/galeri/forgaleri.F90
	packages/ForTrilinos/src/ifpack/forifpack.F90

+++ packages/zoltan/src/lb/lb_remap.c
@@ -120,7 +120,6 @@ End:
   return(ierr);
 }
 
-
 /******************************************************************************/
 static int local_HEs_from_import_lists(
   ZZ *zz,
@@ -595,7 +594,7 @@ float with_oldremap = 0;      /* Amount of data that overlaps between old and
       tmp = i * HEINFO_ENTRIES;
       hg.hindex[i] = i+i; 
       hg.hvertex[i+i] = recvbuf[tmp];
-      hg.hvertex[i+i+1] = recvbuf[tmp+1]+max0;
+      hg.hvertex[i+i+1] = (int)recvbuf[tmp+1]+max0;
       hg.ewgt[i] = recvbuf[tmp+2];
     }
     hg.hindex[total_HEcnt] = total_HEcnt + total_HEcnt;
@@ -731,7 +730,7 @@ int tmp, i;
   for (i = 0; i < hg->nEdge; i++) {
     tmp = i + i;
     if (remapvec) {
-      if (hg->hvertex[tmp] == remapvec[hg->hvertex[tmp+1]-max0]) 
+      if (hg->hvertex[tmp] == (int)remapvec[hg->hvertex[tmp+1]-max0]) 
         stay += hg->ewgt[i];
     }
     else {
